+++ CHANGELOG.md
@@ -14,11 +14,13 @@
 * [Nix/Utilities] Add `dnsutils` and `smartmontools`
 * [Zsh/Aliases] Add the `random-string <num_chars>` function
 * [Zsh/Git] Add setup and aliases for `hub`, the GitHub CLI
+* [Zsh/Nix] Add aliases to diff the package list before (configuration) updates
 
 ### Enhancements
 
 * [Nix/Environment] Create the `$TMPDIR` before to rebuild in `nors`
 * [Nix/Utilities] Add `inxi`, `lshw`, `pciutils` and `usbutils`
+* [Nix/Utilities] Add `jq` and `colordiff`
 * [ranger/bepo] Add `È` to edit as root
 * [ranger/bepo] Add `gt` to go to `/tmp`
 * [Tmux] Use the color 231 instead of 15 for white to avoid issue with terminal